Moved Role.java to test sources
